added random agent for testing - does basic goal card play and hand emptying
